Fix imports after move of live_image payload
