+++ pyanaconda/modules/payload/payloads/live_image/live_image.py
@@ -30,7 +30,8 @@ from pyanaconda.modules.common.constants.objects import LIVE_IMAGE_HANDLER
 from pyanaconda.modules.common.errors.payload import SourceSetupError
 from pyanaconda.modules.payload.base.handler_base import PayloadHandlerBase
 from pyanaconda.modules.payload.base.initialization import CopyDriverDisksFilesTask
-from pyanaconda.modules.payload.live.live_image_interface import LiveImageHandlerInterface
+from pyanaconda.modules.payload.payloads.live_image.live_image_interface import \
+    LiveImageHandlerInterface
 from pyanaconda.modules.payload.live.initialization import CheckInstallationSourceImageTask, \
     SetupInstallationSourceImageTask, UpdateBLSConfigurationTask, \
     TeardownInstallationSourceImageTask
